\name{adam-package}
\alias{adam-package}
\alias{adam}
\docType{package}
\title{
\packageTitle{adam}
}
\description{
\packageDescription{adam}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{adam}
\packageIndices{adam}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{adam}

Maintainer: \packageMaintainer{adam}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
