\name{ADAM.empiricalOdds}
\alias{ADAM.empiricalOdds}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Empirical odds of number of fitness genes per number of cell lines
}
\description{
This function calculates log10 odd ratios of observed vs. expected profiles of cumulative number of fitness genes in fixed number of cell lines. Expected values are the mean of those observed across randomised version of the observed binary matrix.
}
\usage{
ADAM.empiricalOdds(observedCumSum,simulatedCumSum)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{observedCumSum}{
Observed profile of cumulative sum of numbers of fitness genes in fixed number of cell lines. This is generated by the \code{ADAM.panessprofile} function.
}
  \item{simulatedCumSum}{
Random profiles of cumulative sum of fitness genes in fixed number of cell lines. This is generated by the function \code{ADAM.generateNullModel}.
  }
}
\details{
}
\value{
A named vector:
\item{odds}{log base 10 odd ratios of observed versus expected cumulative sums of number of fitness genes across fixed numbers of cell lines.}
}
\references{
}
\author{
C. Pacini & F. Iorio
}
\note{
%%  ~~further notes~~
}

\seealso{
ADAM.panessprofile, ADAM.generateNullModel
}
\examples{

data(exampleDepMat)
observed<-ADAM.panessprofile(depMat=exampleDepMat)
null_m<-ADAM.generateNullModel(depMat=exampleDepMat)
logOdds <- ADAM.empiricalOdds(observed$CUMsums,null_m$nullCumSUM)

logOdds
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
}
\keyword{functions}% use one of  RShowDoc("KEYWORDS")
