% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADAM_main.R
\name{ADAM_main}
\alias{ADAM_main}
\title{ADAM - Adaptive Thresholds Main Function}
\usage{
ADAM_main(
  data,
  norm_method = "TSS",
  perc_strata = 0.2,
  OTU_replace = TRUE,
  threshold_value = TRUE,
  alpha = 0.05,
  B = 1,
  path = getwd(),
  dataset = 1
)
}
\arguments{
\item{data}{A dataset containing bulk simulation data or OTU table to analyze.}

\item{norm_method}{Normalization method to use. Default is 'TSS'.}

\item{perc_strata}{Proportion of strata for breaking data. Default is 0.2.}

\item{OTU_replace}{Logical indicating whether to replace OTUs during training data generation. Default is TRUE.}

\item{threshold_value}{Logical indicating whether to use thresholding in the analysis. Default is TRUE.}

\item{alpha}{Significance level for FDR control. Default is 0.05.}

\item{B}{Number of bootstrap/permutation datasets. Default is 2.}

\item{path}{File path to save results. Default is the current working directory.}

\item{dataset}{A numeric identifier for the dataset. Default is 1.}
}
\value{
A list containing the final results of the ADAM workflow.
}
\description{
This is the main function of the ADAM (Adaptive Thresholds) workflow. It performs preprocessing,
Test data creation, bootstrapping, method selection, and final analysis to identify the
best statistical method for differential abundance analysis.
}
\examples{
\dontrun{
  library(dplyr)
  path <- getwd()
  dataset <- 1
  perc_strata <- 0.2
  OTU_replace <- TRUE
  norm_method <- 'TSS'
  B <- 2
  threshold_value <- TRUE
  alpha <- 0.05
  data <- sim.data.bulk.p_B[[1]]
  
  final_results <- ADAM_main(
    data = data,
    norm_method = norm_method,
    perc_strata = perc_strata,
    OTU_replace = OTU_replace,
    threshold_value = threshold_value,
    alpha = alpha,
    B = B,
    path = path,
    dataset = dataset
  )
}

}
