% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aldex_method.R
\name{aldex_method}
\alias{aldex_method}
\title{ALDEx2 Method}
\usage{
aldex_method(data)
}
\arguments{
\item{data}{A phyloseq object containing count data and metadata.}
}
\value{
A list containing t-test and Wilcoxon test results, as well as the full ALDEx2 output.
}
\description{
This function applies the ALDEx2 method for differential abundance analysis using compositional data.
}
\examples{
\dontrun{
  results <- aldex_method(data)
}

}
