% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANCOMBC_method.R
\name{ANCOMBC_method}
\alias{ANCOMBC_method}
\title{ANCOM-BC Method}
\usage{
ANCOMBC_method(data)
}
\arguments{
\item{data}{A phyloseq object containing count data and metadata.}
}
\value{
A data frame with test statistics, p-values, and adjusted p-values.
}
\description{
This function applies the ANCOM-BC method for differential abundance analysis.
}
\examples{
\dontrun{
  results <- ANCOMBC_method(data)
}

}
