% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_methods_stat.R
\name{apply_methods_stat}
\alias{apply_methods_stat}
\title{apply_methods_stat: Statistical Method Application Function}
\usage{
apply_methods_stat(data)
}
\arguments{
\item{data}{A dataset on which to apply the statistical methods.}
}
\value{
A list containing results from various statistical methods.
}
\description{
This function applies multiple statistical methods to a given dataset and returns
the results in a structured list. It handles errors gracefully by using \code{tryCatch}.
}
\examples{
\dontrun{
  results <- apply_methods_stat(data)
}
}
