% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{AvailableDBSpecies}
\alias{AvailableDBSpecies}
\title{Available species annotation packages}
\format{A data.frame with 19 rows and 3 variables}
\description{
Contains a data.frame describing all available species
annotation packages from Bioconductor.
}
\examples{
data(AvailableDBSpecies)
}
\references{
GENTLEMAN, Robert C. et al. Bioconductor: open software 
development for computational biology and bioinformatics. Genome biology,
v. 5, n. 10, p. R80, 2004.
}
