% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_original_data.R
\name{best_method}
\alias{best_method}
\title{Analyze Original Data}
\usage{
best_method(data, method)
}
\arguments{
\item{data}{A phyloseq object containing the original dataset.}

\item{method}{The statistical method to apply, as a string.}

\item{res_best_method}{A data frame containing the results of the best-performing method,
including metrics such as Type I error rate, sensitivity, and FDR.}
}
\value{
A list containing:
\describe{
\item{\code{final_res}}{Performance metrics (Type I error rate, sensitivity, FDR).}
\item{\code{method}}{The statistical method used.}
}
}
\description{
This function analyzes the original data using the best-performing method, evaluates the results,
and provides performance metrics such as Type I error rate, sensitivity, and FDR.
}
\examples{
\dontrun{
  results <- analyze_original_data(data, method = "DESeq2", res_best_method = res_best_method)
}

}
