% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_best_method.R
\name{choose_best_method}
\alias{choose_best_method}
\title{Choose Best Method}
\usage{
choose_best_method(data_training, res_intermediate, alpha, B, k, path)
}
\arguments{
\item{data_training}{A phyloseq object containing training data.}

\item{res_intermediate}{A list of results from intermediate analysis using different methods.}

\item{B}{Number of bootstrap/permutation datasets.}

\item{k}{An identifier for the training dataset.}

\item{path}{Directory path to save the results.}
}
\value{
A list containing:
\describe{
\item{\code{res_all}}{Summary metrics for all methods.}
\item{\code{res_best}}{The best-performing method based on the evaluation criteria.}
\item{\code{res_all_tmp}}{Detailed results for all methods.}
}
}
\description{
This function evaluates multiple statistical methods based on their performance on permutation/bootstrap datasets.
It calculates metrics such as Type I error rate, sensitivity, and FDR, and selects the best-performing method.
}
\examples{
\dontrun{
  best_method <- choose_best_method(data_training, res_intermediate, B = 100, k = 1, path = "results/")
}

}
