% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.r
\name{collapse_rows}
\alias{collapse_rows}
\title{Collapse the rows of a data.frame object}
\usage{
collapse_rows(x, key = "USUBJID", collapse_name = "data")
}
\description{
Collapse the rows of a data.frame object
}
\examples{
data(adorirr)
collapse_rows(adorirr)
}
