% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.r
\name{contradicting_vars}
\alias{contradicting_vars}
\title{Find the Data Sets with Conflicting Columns}
\usage{
contradicting_vars(x, on)
}
\arguments{
\item{x}{the list of ADaM data sets.}

\item{on}{which variable should be collpased on?}
}
\description{
Find the Data Sets with Conflicting Columns
}
