% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.r
\name{contradiction_tibbles}
\alias{contradiction_tibbles}
\title{Create a Set of data.frame with Contradictions.}
\usage{
contradiction_tibbles(x, on)
}
\arguments{
\item{x}{the list of ADaM data sets.}

\item{on}{which variable should be collpased on?}
}
\description{
Create a Set of data.frame with Contradictions.
}
