% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_training_data.R
\name{create_training_data}
\alias{create_training_data}
\title{Create Test Dataset for Microbiome Analysis}
\usage{
create_training_data(
  data,
  TPR,
  TPR_B,
  perc_strata,
  norm_method = "TSS",
  threshold_I1,
  threshold_I01,
  threshold_I02,
  OTU_replace,
  combinations,
  combinations_I0,
  combinations_I1,
  combinations_B,
  path,
  k
)
}
\arguments{
\item{data}{A phyloseq object containing the microbiome data.}

\item{TPR}{True Positive Rate for the I1 group (numeric).}

\item{TPR_B}{True Positive Rate for the B group (numeric).}

\item{perc_strata}{Sampling fraction (numeric).}

\item{OTU_replace}{Boolean indicating whether to replace sampled OTUs.}

\item{combinations}{Data frame of combinations (data frame).}

\item{combinations_I0}{Data frame of I0 combinations (data frame).}

\item{combinations_I1}{Data frame of I1 combinations (data frame).}

\item{combinations_B}{Data frame of B combinations (data frame).}

\item{path}{Directory path to save training data (character string).}

\item{k}{Index or identifier for the training data (numeric).}

\item{treshold_I1}{Threshold for I1 group (numeric).}

\item{treshold_I01}{Threshold for I0 group (numeric).}

\item{treshold_I02}{Threshold for between groups (numeric).}
}
\value{
A phyloseq object containing the training data.
}
\description{
This function creates test data based on sampling OTUs from different strata.
}
