% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq2_method.R
\name{deseq2_method}
\alias{deseq2_method}
\title{DESeq2 Method}
\usage{
deseq2_method(data)
}
\arguments{
\item{data}{A phyloseq object containing count data and metadata.}
}
\value{
A data frame with DESeq2 results, including test statistics, p-values, and adjusted p-values.
}
\description{
This function applies the DESeq2 method for differential abundance analysis on a given dataset.
}
\examples{
\dontrun{
  results <- deseq2_method(data)
}

}
