% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{DiffHs}
\alias{DiffHs}
\title{Homo sapiens microarray differential expression data example}
\format{A data.frame with 38 rows and 3 variables}
\description{
A sample fragment of differential gene expression from a
microarray experiment of Homo sapiens.
}
\examples{
\dontrun{
data(DiffHs)
}
}
\references{
UEHARA, T. et al. The japanese toxicogenomics project: 
application of toxicogenomics. Molecular nutrition & food research, 
Wiley Online Library, v. 54, n. 2, p. 218–227, 2010.
}
