% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_sums.R
\name{eval_sums}
\alias{eval_sums}
\title{Evaluate Performance Metrics for Statistical Tests}
\usage{
eval_sums(pval, I0_id, I1_id, alpha = 0.05)
}
\arguments{
\item{pval}{A numeric vector of p-values.}

\item{I0_id}{Indices of true negatives (I0).}

\item{I1_id}{Indices of true positives (I1).}

\item{alpha}{Significance level for determining significant p-values. Default is 0.05.}
}
\value{
A numeric vector with three elements:
\describe{
\item{\code{FPP}}{False positive proportion (Type I error rate).}
\item{\code{TPP}}{True positive proportion (sensitivity).}
\item{\code{FDP}}{False discovery proportion.}
}
}
\description{
This function calculates Type I error rate (FPP), sensitivity (TPP),
and false discovery proportion (FDP) based on the provided p-values and
group indices for I0 (true negatives) and I1 (true positives).
}
\examples{
\dontrun{
  pvals <- c(0.01, 0.02, 0.5, 0.04, 0.3)
  I0 <- c(3, 5)
  I1 <- c(1, 2, 4)
  metrics <- eval_sums(pvals, I0, I1, alpha = 0.05)
}

}
