% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ExpressionHs}
\alias{ExpressionHs}
\title{Homo sapiens microarray data expression}
\format{A data.frame with 2000 rows and 5 variables}
\description{
A sample fragment of gene expression from an RNA-seq experiment
of Homo sapiens.
}
\examples{
data(ExpressionHs)
}
\references{
UEHARA, T. et al. The japanese toxicogenomics project: 
application of toxicogenomics. Molecular nutrition & food research, 
Wiley Online Library, v. 54, n. 2, p. 218–227, 2010.
}
