% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{GeneFunctionHs}
\alias{GeneFunctionHs}
\title{Relation between Homo sapiens genes and Cellular Components}
\format{A data.frame with 39 rows and 2 variables}
\description{
A relation between the GFAGs present in ResultAnalysisHs and
their respective genes.
}
\examples{
\dontrun{
data(GeneFunctionHs)
}
}
\references{
Seco, G. B. S. 2018. "Comparative analysis of methods of 
determination of altered biological pathways in toxicogenomic data of in 
vitro and in vivo studies." Master’s thesis, Institute of Biosciences of
Botucatu – Univ. Estadual Paulista. http://hdl.handle.net/11449/157078
}
