% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{KeggPathwaysAedes}
\alias{KeggPathwaysAedes}
\title{Relation between Aedes aegypti genes and KEGG pathways as 
ADAM input}
\format{A data frame with 200 rows and 2 variables}
\description{
A relation between the genes in the ExpressionAedes data and
their respective KEGG pathways (GFAGs).
}
\examples{
data(KeggPathwaysAedes)
}
\references{
Molan, A. L. 2018. “Construction of a Tool for Multispecies
Genic Functional Enrichment Analysis Among Comparative Samples.”
Master’s thesis Institute of Biosciences of Botucatu – Univ. Estadual
Paulista. http://hdl.handle.net/11449/157105.
}
