% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinDA_test.R
\name{LinDA_test}
\alias{LinDA_test}
\title{LinDA Test}
\usage{
LinDA_test(data)
}
\arguments{
\item{data}{A phyloseq object containing count data and metadata.}
}
\value{
A data frame with LinDA test statistics, p-values, and adjusted p-values.
}
\description{
This function applies the LinDA method from the MicrobiomeStat package for differential abundance analysis.
}
\examples{
\dontrun{
  results <- LinDA_test(data)
}

}
