% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorten.r
\name{name_map}
\alias{name_map}
\title{Get the Map of Shortened Names to Original Names}
\usage{
name_map(x)
}
\arguments{
\item{x}{the data.frame to get the name map from.}
}
\value{
if the name map exists, then the return is a tibble where the
column "string" is the original column name and "short_string" is the
shortened column name.
}
\description{
Get the Map of Shortened Names to Original Names
}
