% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.r
\name{normalize_adam}
\alias{normalize_adam}
\title{Normalize an ADam dataset.}
\usage{
normalize_adam(x, on = "USUBJID", collapse_name)
}
\description{
Normalize an ADam dataset.
}
\examples{
data(adorirr)
normalize_adam(adorirr)
}
