% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_comp.R
\name{pairwise_comp}
\alias{pairwise_comp}
\title{Perform Pairwise Comparisons of OTUs}
\usage{
pairwise_comp(
  data,
  break_I0,
  break_I1,
  break_B,
  threshold_I1,
  threshold_I01,
  norm_method = "TSS"
)
}
\arguments{
\item{data}{A phyloseq object containing the OTU data for analysis.}

\item{break_I0}{A data frame containing breakpoints for the I0 strata.}

\item{break_I1}{A data frame containing breakpoints for the I1 strata.}

\item{break_B}{A data frame containing breakpoints for the between strata.}

\item{threshold_I1}{Numeric threshold for high impact OTUs.}

\item{threshold_I01}{Numeric threshold for baseline impact OTUs.}
}
\value{
A list containing:
\item{combinations}{A data frame of pairwise OTU comparisons and their characteristics.}
\item{combinations_I0}{Subset of combinations that fall into the I0 strata.}
\item{combinations_I1}{Subset of combinations that fall into the I1 strata.}
\item{combinations_B}{Subset of combinations that fall into the between strata.}
}
\description{
This function performs pairwise comparisons of OTUs in a microbiome dataset. It calculates various characteristics
of pseudo OTU pairs based on normalized data and stratifies them according to defined thresholds.
}
\examples{
# result <- pairwise_comp(data, break_I0, break_I1, break_B, threshold_I1, threshold_I01)
}
