% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm_boot.R
\name{perm_boot}
\alias{perm_boot}
\title{perm_boot: Permutation and Bootstrap Sampling Function}
\usage{
perm_boot(data_training, B, nSamples, path, dataset)
}
\arguments{
\item{data_training}{Training data for the procedure.}

\item{B}{Number of permutation/bootstrap datasets to generate.}

\item{nSamples}{Number of samples of the original dataset.}

\item{path}{Directory path to save the results (if saving functionality is enabled).}

\item{dataset}{Name of the dataset for identification purposes.}
}
\value{
A list of intermediate results obtained after applying statistical methods.
}
\description{
This function performs permutation and bootstrap sampling on training data
and applies statistical methods to the modified data. The results are saved
in a list.
}
\examples{
\dontrun{
  res <- perm_boot(data_training, B = 100, nSamples = 50, path = "results/", dataset = "example")
}

}
