% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.r
\name{pnc_adam}
\alias{pnc_adam}
\title{Process, Normalize, and, Consolidate ADaM Data Sets}
\usage{
pnc_adam(x, on = "USUBJID", handle_contra_vals = FALSE,
  collapse_name = names(x), remove_equiv_columns = TRUE,
  remove_numerically_encoded = TRUE, keep_cols = character(),
  verbose = TRUE)
}
\arguments{
\item{x}{a named list of data.frames.}

\item{on}{which variable should be collpased on? (Default: "USUBJID")}

\item{handle_contra_vals}{if variables have contradicting values across
data.frame's, then can they be handled by using the first instance of
the variable in the set? (Default FALSE) See Details for more information.}

\item{collapse_name}{the variable name of the collapsed sub-data.frames.
Default is the names of 'x'.}

\item{verbose}{should extra information be provided? (Default TRUE)}
}
\description{
Process, Normalize, and, Consolidate ADaM Data Sets
}
\details{
Setting 'handle_contra_vals' to TRUE should be used with 
extreme caution. If it is used then the contradicting value are stored
as an attribute of the consolidated data.frame and can be retrieved using
the function 'get_contradictions()'.
}
