% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-adam-description.r
\name{report_adam_desc}
\alias{report_adam_desc}
\title{Generate an ADaM Description Report}
\usage{
report_adam_desc(x, title = "ADaM Data Description",
  author = "Some Person", date = as.character(Sys.Date()),
  output = paste0(tempfile(), ".html"), keep_artifacts = FALSE,
  cleanup = FALSE, launch_browser = interactive())
}
\arguments{
\item{x}{the data frame to generate the report for.}

\item{title}{the title of the document. (Default "Missingness Report")}

\item{author}{the name of the author of the document.
(Default "Some Person")}

\item{date}{the date associated with the output document.
(Default is the return of as.character(Sys.Date()))}

\item{output}{the name of the output file, ending in .html, .pdf, or .docx.
(Default it a temporary .html file)}

\item{keep_artifacts}{should the data and .rmd files be kept? (default TRUE)}

\item{cleanup}{should temporary files/directories be deleted? (default TRUE)}

\item{launch_browser}{should a browser be launched to examine the report?
(default is TRUE for an interactive session, FALSE otherwise)}
}
\description{
Generate a report giving a short description of the
missingness in a data set. The report can be rendered in html, word,
or pdf and includes a plot of the number of missing values by variable,
an upset plot showing the co-occurence of missing values, and a
missing variable summary.
}
