% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorten.r
\name{shorten_acronym}
\alias{shorten_acronym}
\title{Shorten Character Strings to Acronym}
\usage{
shorten_acronym(x, sep = " ", cruft_pattern = "[^a-zA-Z0-9 \\\\.]")
}
\arguments{
\item{x}{a vector of characters.}

\item{sep}{the string separators. Default " ".}

\item{cruft_pattern}{the regex desribing the characters that will be dropped
from the string.}
}
\description{
Shorten Character Strings to Acronym
}
