% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorten.r
\name{shorten_string}
\alias{shorten_string}
\title{Shorten Character Strings}
\usage{
shorten_string(x, min_len = 3, max_len = 10, max_total = 18,
  sep = " ", outsep = "_", cruft_pattern = "[aeiou\\\\-]")
}
\arguments{
\item{x}{a vector of characters.}

\item{min_len}{the minimum length a string can be without being dropped. 
Default 3.}

\item{max_len}{the maximum length a string can be. If it's longer it
gets truncated. Default 10.}

\item{max_total}{the maximum total size of the shortened string.
Default 18.}

\item{sep}{the string separators. Default " ".}

\item{outsep}{the shortened string separator. Default "_".}

\item{cruft_pattern}{the regex desribing the characters that will be dropped
from the string.}
}
\description{
Shorten Character Strings
}
