% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorten.r
\name{shorten_strings}
\alias{shorten_strings}
\title{#export}
\usage{
shorten_strings(x, min_len = 3, max_len = 10, max_total = 18,
  sep = " ", outsep = "_", cruft_pattern = "[aeiou\\\\-]")
}
\description{
#export
}
