% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtract_info.R
\name{subtract_info}
\alias{subtract_info}
\title{Subtract information from dataset based on thresholds}
\usage{
subtract_info(
  data,
  threshold_I1 = NULL,
  threshold_I01 = NULL,
  threshold_I02 = NULL,
  threshold_value = FALSE,
  norm_method = "TSS",
  log_file = "test_log.txt",
  perc_strata = 0.1
)
}
\arguments{
\item{data}{A phyloseq object containing OTU data for analysis.}

\item{threshold_I1}{Numeric threshold for I1. If not provided, calculated from log fold changes.}

\item{threshold_I01}{Numeric threshold for I01. If not provided, calculated from log fold changes.}

\item{threshold_I02}{Numeric threshold for I02. If not provided, calculated from log fold changes.}

\item{threshold_value}{Logical indicating whether to compute the thresholds from the data (default is FALSE).}

\item{norm_method}{Character, either "TSS" or "CLR" to specify the normalization method (default: "TSS").}

\item{log_file}{Path to a log file where data-driven thresholds are logged (default: "log.txt").}

\item{perc_strata}{Numeric, indicating the fraction of data to be used per strata (default: 0.1).}
}
\value{
A list with breakpoints for the different strata (I0, I1, between), TPR values, sample counts, and thresholds.
}
\description{
This function processes a microbiome dataset by normalizing the data using CLR (centered log-ratio) or TSS (Total Sum Scaling) transformation.
It calculates log fold changes (lfc) between two groups and applies thresholds to determine true positive rates (TPR).
The function then creates strata for the OTUs based on relative abundances and thresholds.
}
\examples{
data <- example_data()  # Your data loading code here
result <- subtract_info(data)
}
