% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.r
\name{tcat}
\alias{tcat}
\title{Print a message}
\usage{
tcat(x, msg, verbose = TRUE, style = reset, ...)
}
\arguments{
\item{x}{the data the function will return}

\item{msg}{the message to cat.}

\item{verbose}{should the message be cat'ed? Default TRUE.}

\item{style}{the crayon style to use when printing. Default reset.}

\item{...}{other paramters passed to cat.}
}
\description{
Print a message
}
