% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.r
\name{var_equiv_violation}
\alias{var_equiv_violation}
\title{Test for Variable Equivence in a list of data.frames}
\usage{
var_equiv_violation(x, on)
}
\arguments{
\item{x}{the list of data.frames.}

\item{on}{the collapse variable.}
}
\description{
Test for Variable Equivence in a list of data.frames
}
