% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.r
\name{collapse_rows}
\alias{collapse_rows}
\title{Collapse the rows of a data.frame object}
\usage{
collapse_rows(x, key = "USUBJID", collapse_name = "data")
}
\arguments{
\item{x}{and ADaM formatted data.frame}

\item{key}{which variable should be collpased on? (Default: "USUBJID")}

\item{collapse_name}{the variable name of the collapsed sub-data.frames.}
}
\description{
Collapse the rows of a data.frame object
}
