% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.r
\name{consolidate_adam}
\alias{consolidate_adam}
\title{Consolidate multiple data sets}
\usage{
consolidate_adam(..., on = "USUBJID", verbose = FALSE)
}
\arguments{
\item{...}{a set of ADaM formatted data.frames.}

\item{on}{which variable should be collpased on? (Default: "USUBJID")}

\item{verbose}{should extra information be provided? (Default: TRUE)}
}
\value{
A single data.frame composed of the collapsed and merged input
data.frames.
}
\description{
Consolidate multiple data sets
}
