% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.r
\name{consolidate_adam}
\alias{consolidate_adam}
\title{Consolidate multiple data sets}
\usage{
consolidate_adam(..., on = "USUBJID")
}
\description{
Consolidate multiple data sets
}
\examples{
data(adorirr, adefirr, adsl)
ruca <- list(normalize_adam(adorirr), normalize_adam(adefirr), 
             normalize_adam(adsl))
x <- consolidate_adam(ruca)
}
