% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.r
\name{normalize_adam}
\alias{normalize_adam}
\title{Normalize an ADam dataset.}
\usage{
normalize_adam(x, on = "USUBJID", collapse_name,
  remove_equiv_columns = TRUE, remove_numerically_encoded = TRUE,
  keep_cols = character(), verbose = FALSE)
}
\arguments{
\item{x}{an ADaM formatted data.frame.}

\item{on}{which variable should be collpased on? (Default: "USUBJID")}

\item{collapse_name}{the variable name of the collapsed sub-data.frames.}

\item{remove_equiv_columns}{should equivalent columns be removed?}

\item{remove_numerically_encoded}{should numerically encoded columns be 
removed? Default TRUE}

\item{keep_cols}{column names that should be kept.}

\item{verbose}{should information about dropped columns be printed? 
(default FALSE)}
}
\value{
The collapsed data.frame with numerically encoded columens removed.
}
\description{
Normalize an ADam dataset.
}
