% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.r
\name{normalize_adam}
\alias{normalize_adam}
\title{Normalize an ADam dataset.}
\usage{
normalize_adam(x, on = "USUBJID", collapse_name)
}
\arguments{
\item{x}{an ADaM formatted data.frame.}

\item{on}{which variable should be collpased on? (Default: "USUBJID")}

\item{collapse_name}{the variable name of the collapsed sub-data.frames.}
}
\value{
The collapsed data.frame with numerically encoded columens removed.
}
\description{
Normalize an ADam dataset.
}
\examples{
data(adorirr)
normalize_adam(adorirr)
}
