% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.r
\name{remove_numerically_encoded_columns}
\alias{remove_numerically_encoded_columns}
\title{Remove numerically encoded columns}
\usage{
remove_numerically_encoded_columns(x)
}
\arguments{
\item{x}{the ADaM formatted data.frame which may have numerically encoded
columns.}
}
\description{
Remove numerically encoded columns
}
\examples{
data(adorirr)
remove_numerically_encoded_columns(adorirr)
}
