% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.r
\name{repeat_vars}
\alias{repeat_vars}
\title{Find Variables Appearing Multiple Data Sets}
\usage{
repeat_vars(x, on, x_names = names(x))
}
\arguments{
\item{x}{the list of data.frames.}

\item{on}{the collapse variable.}

\item{remove_nas}{remove NA before testing for equivalence? Note all values
at all NA positions are removed.}
}
\description{
Find Variables Appearing Multiple Data Sets
}
